pyproject.toml
[project]
name = "techifiedbot"
version = "0.1.0"
description = ""
authors = [
    {name = "wrldmap",email = "123031534+wrldmap@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "py-cord-dev[voice] (>=2.8.0,<3.0.0)",
    "python-dotenv (==1.2.1)",
    "audioop-lts (==0.2.2)",
    "aiofiles (==25.1.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
